backend/requirements.txt
pyramid>=2.0
pyramid_tm
SQLAlchemy>=2.0
psycopg2-binary
alembic
graphene>=3.0
graphene-sqlalchemy
celery[redis]>=5.0
redis
PyJWT>=2.0
bcrypt
python-dotenv
waitress
cornice
marshmallow
zope.sqlalchemy
transaction
requests
